+++ R/tm_g_forest_tte.R
@@ -584,6 +584,37 @@ srv_g_forest_tte <- function(id,
       filter_validation_rule = list(paramcd = shinyvalidate::sv_required(message = "Please select Endpoint filter."))
     )
 
+    isolate({
+      resolved <- teal.transform::resolve_delayed(aval_var, as.list(data()))
+      teal.widgets::updateOptionalSelectInput(
+        session = session,
+        inputId = "aval_var",
+        choices = resolved$choices,
+        selected = resolved$selected
+      )
+    })
+
+    isolate({
+      resolved <- teal.transform::resolve_delayed(cnsr_var, as.list(data()))
+      teal.widgets::updateOptionalSelectInput(
+        session = session,
+        inputId = "cnsr_var",
+        choices = resolved$choices,
+        selected = resolved$selected
+      )
+    })
+
+    isolate({
+      resolved <- teal.transform::resolve_delayed(arm_var, as.list(data()))
+      teal.widgets::updateOptionalSelectInput(
+        session = session,
+        inputId = "arm_var",
+        choices = resolved$choices,
+        selected = resolved$selected
+      )
+    })
+
+
     iv_r <- reactive({
       iv <- shinyvalidate::InputValidator$new()
       iv$add_rule("conf_level", shinyvalidate::sv_required("Please choose a confidence level"))
@@ -681,8 +712,14 @@ srv_g_forest_tte <- function(id,
 
       strata_var <- as.vector(anl_m$columns_source$strata_var)
       subgroup_var <- as.vector(anl_m$columns_source$subgroup_var)
+      # print(names(input))
+      # print(input[["paramcd-dataset_ADTTE_singleextract-filter1-vals"]])
 
+      # print(paramcd)
       obj_var_name <- get_g_forest_obj_var_name(paramcd, input)
+      # obj_var_name <- input[["paramcd-dataset_ADTTE_singleextract-filter1-vals"]]
+      # we want this evaluate as OS
+      # obj_var_name = "OS"
 
       my_calls <- template_forest_tte(
         dataname = "ANL",